requirements.txt
keras
tensorflow
sounddevice
googletranslator
numpy
